precalculated distance between squares

+++ include/evaluation/evaluation.hpp
@@ -37,4 +37,4 @@ constexpr int32_t MATE_IN_ONE_SCORE = 32000U;
  *  - (+) if position is evaluated as white is better.
  *  - (-) if position is evaluated as black is better.
  */
-float evaluate_position(const Board& board);
+int evaluate_position(const Board& board);
